affichage ptn

+++ headers/struct.h
@@ -6,7 +6,7 @@
 /*   By: astachni <astachni@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 17:15:02 by astachni          #+#    #+#             */
-/*   Updated: 2023/10/04 21:59:05 by astachni         ###   ########.fr       */
+/*   Updated: 2023/10/05 18:24:09 by astachni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct s_fpoint
 	float	x;
 	float	y;
 	float	distance;
-	int		height;
+	float	height;
 }	t_fpoint;
 
 typedef struct s_point
